todo test inputstream.

+++ executor/src/main/java/cn/jdworks/etl/executor/biz/TaskRunner.java
@@ -11,55 +11,71 @@ import org.nutz.log.Log;
 import org.nutz.log.Logs;
 
 public class TaskRunner extends Thread {
-	public void startRunner(int id, String cmd, TaskEventHandler eventHandler) {
+
+	public synchronized boolean startRunner(int id, String cmd, TaskEventHandler eventHandler) {
 		this.eventHandler = eventHandler;
 		this.id = id;
 		this.command = cmd;
 
-		if (runTask()){
+		if (runTask()) {
 			this.setRunning(true);
 			this.start();
+			return true;
+		} else {
+			return false;
 		}
 	}
-	
-	public void abortTask() {
-		this.taskProc.destroyForcibly();
-		this.setRunning(false);
-		try {
-			this.join();
-		} catch (Throwable t) {
-			LOG.debug(t);
+
+	public synchronized void abortTask() {
+		if (isRunning()) {
+			this.taskProc.destroyForcibly();
+			this.setRunning(false);
+			try {
+				this.join();
+			} catch (Throwable t) {
+				LOG.debug(t);
+			}
 		}
 	}
-	
+
 	private final Log LOG = Logs.getLog(this.getClass());
 
 	private TaskEventHandler eventHandler = null;
-	private void logStarted(){
-		if(this.eventHandler !=null)
+
+	private void logStarted() {
+		if (this.eventHandler != null)
 			this.eventHandler.onTaskStarted(id, System.currentTimeMillis());
 	}
-	private void logStopped(int exit){
-		if(this.eventHandler !=null)
+
+	private void logStopped(int exit) {
+		if (this.eventHandler != null)
 			this.eventHandler.onTaskStopped(id, System.currentTimeMillis(), exit);
 	}
-	private void logStartFailed(String error){
-		if(this.eventHandler !=null)
+
+	private void logStartFailed(String error) {
+		if (this.eventHandler != null)
 			this.eventHandler.onTaskStartFailed(id, System.currentTimeMillis(), error);
 	}
-	private void logged(long ts, String type, String message){
-		if(this.eventHandler !=null)
+
+	private void logged(long ts, String type, String message) {
+		if (this.eventHandler != null)
 			this.eventHandler.onTaskLogged(id, ts, type, message);
 	}
 
+	private void errorLogged(long ts, String error) {
+		if (this.eventHandler != null)
+			this.eventHandler.onTaskErrorLogged(id, ts, error);
+	}
 
 	private Process taskProc;
 	private BufferedReader stdoutReader;
 	private BufferedReader stderrReader;
 	private boolean isRunning = false;
+
 	private synchronized boolean isRunning() {
 		return isRunning;
 	}
+
 	private synchronized void setRunning(boolean isRunning) {
 		this.isRunning = isRunning;
 	}
@@ -67,14 +83,14 @@ public class TaskRunner extends Thread {
 	private int id;
 	private String command;
 
-
 	private boolean runTask() {
 		try {
 			ProcessBuilder pb = new ProcessBuilder();
 			pb.command(command.split(" "));
 			this.taskProc = pb.start();
-			this.start();
 
+			InputStreamReader r = new InputStreamReader(taskProc.getInputStream());
+			rrr///todo...test r.
 			this.stdoutReader = new BufferedReader(new InputStreamReader(taskProc.getInputStream()));
 			this.stderrReader = new BufferedReader(new InputStreamReader(taskProc.getErrorStream()));
 
@@ -96,41 +112,57 @@ public class TaskRunner extends Thread {
 		String message;
 		while (isRunning()) {
 			try {
-				isReady = false;
-
-				if (stdoutReader.ready()) {
-					isReady = true;
-					line = stdoutReader.readLine();
-					if(line==null)
-						throw new Exception();
-					//Log format: 
-					//    ts:type:message
-					try {
-						ss = line.split(":");
-						ts = Long.parseLong(ss[0]);
-						type = ss[1];
-						message = ss[2];
-					} catch (Throwable t1) {
-						ts = System.currentTimeMillis();
-						type="debug";
-						message = line;
-					}
-					this.logged(ts, type, message);
-				}
+				line = null;
+				LOG.debug("*************");
 				if (stderrReader.ready()) {
-					isReady = true;
+					LOG.debug("=============");
 					line = stderrReader.readLine();
-					if(line==null)
-						throw new Exception();
-					this.logged(System.currentTimeMillis(), "error", line);
+					if (line != null) {
+						this.errorLogged(System.currentTimeMillis(), line);
+					}
 				}
-				if (!isReady)
-					Thread.sleep(100);
+				if (stdoutReader.ready()) {
+					LOG.debug("-------------");
+					line = stdoutReader.readLine();
+					if (line != null) {
+						// Log format:
+						// ts:type:message
+						try {
+							ss = line.split(":");
+							ts = Long.parseLong(ss[0]);
+							type = ss[1];
+							message = ss[2];
+						} catch (Throwable t1) {
+							ts = System.currentTimeMillis();
+							type = "debug";
+							message = line;
+						}
+						this.logged(ts, type, message);
+					}
+				}
+
+				if (!stdoutReader.ready() && !stderrReader.ready()) {
+					if (this.taskProc.isAlive())
+						Thread.sleep(100);
+					else
+						throw new Exception("task is destroyed with no more output, let's quit.");
+				} else if (line == null) {
+					throw new Exception("data is ready for reading but read returns NULL.");
+				}
 			} catch (Throwable t) {
 				this.setRunning(false);
 			}
 		}
-		int exit = this.taskProc.exitValue();
+
+		int exit;
+		try {
+			//We may not wait for task exit to quit, so it is possible that here task is still alive
+			//We give this task -1 exit value. 
+			//This does not mean task can not be terminated itself, only means we can not wait for its termination.
+			exit = this.taskProc.exitValue();
+		} catch (Exception e) {
+			exit = -1;
+		}
 		this.logStopped(exit);
 	}
 }
